requirements.txt
face_recognition
numpy
sklearn
Pillow
pickle